+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - 新增 telqos 指令。
   - com.dwarfeng.settingrepo.impl.service.telqos.SettingNodeCommand。
+  - com.dwarfeng.settingrepo.impl.service.telqos.ImageNodeCommand。
 
 - 增加预设的运维指令。
   - com.dwarfeng.springtelqos.api.integration.log4j2.Log4j2Command。